requirements.txt
flask==3.0.3
flask-cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
